[FLINK-26789][tests] Fix broken RescaleCheckpointManuallyITCase
